Temporarily disable calendar
In AddEditSheet, this import was apparently not used at all, and yet somehow both causing a crash on launch due to QtWebEngine not being initialized properly.
Even if QtWebEngine was initialized explicitly in main, this was preventing the calendar from appearing.
I don't know how one line of unneeded code caused so many problems.

This also occurred with other instances where the calendar was used, but those were harder to discover from the perspective of the user.
It has all been disabled until a better solution can be found.