Fix color picker dialog appearing below add/edit sheet
The OverlaySheet is now higher in the rendering order than the main window, and ColorDialog is outside of this hierarchy, so it gets drawn below the overlay sheet and can't actually be seen.
I suspect that this was due to the OverlayZStacking change in KDE plasma 6.
The window hack is suboptimal, and better solutions would be great.